requirements.txt
numpy
scikit-surprise
